Dockerdev: Allow using docker container locally
The current local dev container might not be appropriate for everyone to
do local, vscode-less development. It is somewhat married to the
`.devcontainer` setup, which is showing its warts a little bit already,
where the existing MS container is being modified to more closely match
our desires.

This can easily be fixed by using the normal python container for local
builds, making the runtime container and dev containers closer, which is
always desirable.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>