Add turn_on service to test_service_descriptions
The turn_on service is a built-in service, but it was not being included in
the test_service_descriptions test. This has been fixed by adding the turn_on
service to the test.

Fixes assertion error:
        # Test 4: verify that names from YAML are taken into account as well
>       assert descriptions[DOMAIN]["turn_on"]["name"] == "Turn on"
E       AssertionError: assert '' == 'Turn on'
E         - Turn on

tests/components/script/test_init.py:561: AssertionError

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>